+++ src/dotnet/RiderPlugin.UnrealLink/PluginInstaller/UnrealPluginInstallInfo.cs
@@ -18,7 +18,7 @@ namespace RiderPlugin.UnrealLink.PluginInstaller
 
         public VirtualFileSystemPath EngineRoot = null;
 
-        public InstallDescription EnginePlugin = new InstallDescription();
+        public InstallDescription EnginePlugin = null;
         public List<InstallDescription> ProjectPlugins = new List<InstallDescription>();
 
         public PluginInstallLocation Location;