+++ src/rider/main/kotlin/com/jetbrains/rider/plugins/unreal/notifications/OutOfSyncEditorNotification.kt
@@ -8,25 +8,31 @@ import com.jetbrains.rider.UnrealLinkBundle
 import com.jetbrains.rider.plugins.unreal.model.frontendBackend.*
 import com.jetbrains.rider.projectView.solution
 
-class OutOfSyncEditorNotification(project: Project): ProtocolSubscribedProjectComponent(project) {
+class OutOfSyncEditorNotification(project: Project) : ProtocolSubscribedProjectComponent(project) {
     // TODO: move all public strings to resource bundle
     companion object {
         private val notificationGroupId =
-                NotificationGroup.createIdWithTitle("Unreal Editor connection is out of sync",
-                        UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.groupId"))
+            NotificationGroup.createIdWithTitle(
+                "Unreal Editor connection is out of sync",
+                UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.groupId")
+            )
 
     }
 
     init {
         project.solution.rdRiderModel.onEditorPluginOutOfSync.adviseNotNull(projectComponentLifetime) {
-            if(it.status == PluginInstallStatus.UpToDate) return@adviseNotNull
+            if (it.status == PluginInstallStatus.UpToDate) return@adviseNotNull
 
-            val message = if(it.status == PluginInstallStatus.NoPlugin)
+            val message = if (it.status == PluginInstallStatus.NoPlugin)
                 UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.description.notInstalled")
             else
-                UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.description.wrongVersion", it.installedVersion, it.requiredVersion)
+                UnrealLinkBundle.message(
+                    "notificationAction.UnrealEditorOutOfSync.description.wrongVersion",
+                    it.installedVersion,
+                    it.requiredVersion
+                )
 
-            val title = if(it.status == PluginInstallStatus.NoPlugin)
+            val title = if (it.status == PluginInstallStatus.NoPlugin)
                 UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.title.notInstalled")
             else
                 UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.title.wrongVersion")
@@ -37,13 +43,13 @@ class OutOfSyncEditorNotification(project: Project): ProtocolSubscribedProjectCo
             when (it.status) {
                 PluginInstallStatus.NoPlugin -> {
                     notification.apply {
-                        addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInEngine.text")) {
+                        if (it.isEngineAvailable) addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInEngine.text")) {
                             expire()
                             project.solution.rdRiderModel.installEditorPlugin.fire(
                                 InstallPluginDescription(PluginInstallLocation.Engine, ForceInstall.No)
                             )
                         })
-                        if(it.isGameAvailable) addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInGame.text")) {
+                        if (it.isGameAvailable) addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInGame.text")) {
                             expire()
                             project.solution.rdRiderModel.installEditorPlugin.fire(
                                 InstallPluginDescription(PluginInstallLocation.Game, ForceInstall.No)
@@ -53,13 +59,13 @@ class OutOfSyncEditorNotification(project: Project): ProtocolSubscribedProjectCo
                 }
                 PluginInstallStatus.InEngine -> {
                     notification.apply {
-                        addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInEngine.text.update")) {
+                        if (it.isEngineAvailable) addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInEngine.text.update")) {
                             expire()
                             project.solution.rdRiderModel.installEditorPlugin.fire(
                                 InstallPluginDescription(PluginInstallLocation.Engine, ForceInstall.No)
                             )
                         })
-                        if(it.isGameAvailable) addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInGame.text.move")) {
+                        if (it.isGameAvailable) addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInGame.text.move")) {
                             expire()
                             project.solution.rdRiderModel.installEditorPlugin.fire(
                                 InstallPluginDescription(PluginInstallLocation.Game, ForceInstall.No)
@@ -69,13 +75,13 @@ class OutOfSyncEditorNotification(project: Project): ProtocolSubscribedProjectCo
                 }
                 PluginInstallStatus.InGame -> {
                     notification.apply {
-                        addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInEngine.text.move")) {
+                        if (it.isEngineAvailable) addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInEngine.text.move")) {
                             expire()
                             project.solution.rdRiderModel.installEditorPlugin.fire(
                                 InstallPluginDescription(PluginInstallLocation.Engine, ForceInstall.No)
                             )
                         })
-                        if(it.isGameAvailable) addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInGame.text.update")) {
+                        if (it.isGameAvailable) addAction(NotificationAction.createSimple(UnrealLinkBundle.message("notificationAction.UnrealEditorOutOfSync.installPluginInGame.text.update")) {
                             expire()
                             project.solution.rdRiderModel.installEditorPlugin.fire(
                                 InstallPluginDescription(PluginInstallLocation.Game, ForceInstall.No)